sale history bug fixed

+++ AgroFamily/Model/IExpensiveModel.cs
@@ -12,7 +12,7 @@ namespace AgroFamily.Model
         public void Add(ExpensiveModel expensive);
         public ExpensiveModel GetById(int id);
         public ObservableCollection<ExpensiveModel> GetByAll();
-        int GetAmountInAMonth(int month, int year);
+        double GetAmountInAMonth(int month, int year);
 
     }
 }
